Бесконечный долбёж рутуба #54
- Статус Error фиксируется при падении загрузки
- Детерминированные ошибки больше не ретраятся
- Авто-транскодирование VP9 перед загрузкой на RuTube
- HDD переведён на общий VideoTranscoder
- Выделен VideoTranscoder в MediaOrcestrator.Modules
