+++ MediaOrcestrator.Rutube/RutubeChannel.cs
@@ -1,6 +1,5 @@
 ﻿using MediaOrcestrator.Modules;
 using Microsoft.Extensions.Logging;
-using System.Diagnostics;
 using System.Globalization;
 using System.Runtime.CompilerServices;
 using System.Text;
@@ -9,7 +8,11 @@ using System.Text.Json;
 namespace MediaOrcestrator.Rutube;
 
 // TODO: Костыль с ILogger<RutubeService>. Желательно сделать полноценную регистрацию модулей в DI.
-public class RutubeChannel(ILogger<RutubeChannel> logger, ILogger<RutubeService> serviceLogger, IToolPathProvider toolPathProvider)
+public class RutubeChannel(
+    ILogger<RutubeChannel> logger,
+    ILogger<RutubeService> serviceLogger,
+    IToolPathProvider toolPathProvider,
+    VideoTranscoder videoTranscoder)
     : ISourceType, IAuthenticatable, IToolConsumer
 {
     public SyncDirection ChannelType => SyncDirection.Full;
@@ -225,17 +228,47 @@ public class RutubeChannel(ILogger<RutubeChannel> logger, ILogger<RutubeService>
         if (string.IsNullOrEmpty(codec))
         {
             logger.LogInformation("Кодек не найден в метаданных, определяем через ffprobe: {FilePath}", filePath);
-            codec = await GetVideoCodecAsync(filePath, cancellationToken);
+            codec = await videoTranscoder.GetVideoCodecAsync(filePath, cancellationToken);
         }
 
         if (string.IsNullOrEmpty(codec))
         {
-            throw new InvalidOperationException("Не удалось определить кодек видео. Убедитесь, что ffprobe доступен.");
+            throw new NonRetriableException("Не удалось определить кодек видео. Убедитесь, что ffprobe доступен.");
         }
 
         if (string.Equals(codec, "vp9", StringComparison.OrdinalIgnoreCase))
         {
-            throw new InvalidOperationException($"Видео с кодеком VP9 не поддерживается RuTube. Файл: {filePath}");
+            logger.LogInformation("Обнаружен VP9, запускаем транскодирование в H.264: {FilePath}", filePath);
+
+            var sourceDir = Path.GetDirectoryName(filePath)
+                            ?? throw new InvalidOperationException($"Не удалось получить папку файла: {filePath}");
+
+            var transcodedPath = Path.Combine(sourceDir, "media.transcoded.mp4");
+
+            var durationStr = media.Metadata?.FirstOrDefault(x => x.Key == "Duration")?.Value;
+            var totalDuration = TimeSpan.Zero;
+            if (durationStr != null && TimeSpan.TryParse(durationStr, out var parsed))
+            {
+                totalDuration = parsed;
+            }
+
+            var transcodeOk = await videoTranscoder.TranscodeVp9ToH264Async(filePath,
+                transcodedPath,
+                totalDuration,
+                null,
+                cancellationToken);
+
+            if (!transcodeOk || !File.Exists(transcodedPath) || new FileInfo(transcodedPath).Length == 0)
+            {
+                throw new NonRetriableException($"Не удалось транскодировать VP9 в H.264 для файла: {filePath}");
+            }
+
+            logger.LogInformation("Транскодирование завершено. Размер: {Size} байт. Новый путь: {Path}",
+                new FileInfo(transcodedPath).Length, transcodedPath);
+
+            filePath = transcodedPath;
+            media.TempDataPath = transcodedPath;
+            codec = "h264";
         }
 
         logger.LogInformation("Кодек видео: {Codec}", codec);
@@ -359,74 +392,6 @@ public class RutubeChannel(ILogger<RutubeChannel> logger, ILogger<RutubeService>
         {
             logger.LogInformation("RuTube: авторизация сохранена в {Path}", result);
             await ui.ShowMessageAsync("Авторизация RuTube сохранена!");
-        }
-    }
-
-    // TODO: Копипаста из HDD. Может пора делать стандартные обёртки в MediaOrcestrator.Modules
-    private async Task<string?> GetVideoCodecAsync(string filePath, CancellationToken cancellationToken)
-    {
-        var ffprobePath = toolPathProvider.GetCompanionPath(WellKnownTools.FFmpeg, "ffprobe");
-
-        if (ffprobePath is null)
-        {
-            return null;
-        }
-
-        try
-        {
-            var psi = new ProcessStartInfo
-            {
-                FileName = ffprobePath,
-                Arguments = $"-v quiet -print_format json -show_streams \"{filePath}\"",
-                RedirectStandardOutput = true,
-                RedirectStandardError = true,
-                UseShellExecute = false,
-                CreateNoWindow = true,
-            };
-
-            using var process = Process.Start(psi);
-
-            if (process is null)
-            {
-                return null;
-            }
-
-            var output = await process.StandardOutput.ReadToEndAsync(cancellationToken);
-            await process.WaitForExitAsync(cancellationToken);
-
-            if (process.ExitCode != 0)
-            {
-                logger.LogWarning("ffprobe завершился с кодом {ExitCode} для файла: {FilePath}", process.ExitCode, filePath);
-                return null;
-            }
-
-            using var doc = JsonDocument.Parse(output);
-
-            if (!doc.RootElement.TryGetProperty("streams", out var streams))
-            {
-                return null;
-            }
-
-            foreach (var stream in streams.EnumerateArray())
-            {
-                var codecType = stream.TryGetProperty("codec_type", out var ct) ? ct.GetString() : null;
-
-                if (codecType == "video" && stream.TryGetProperty("codec_name", out var codecName))
-                {
-                    return codecName.GetString();
-                }
-            }
-
-            return null;
-        }
-        catch (OperationCanceledException)
-        {
-            throw;
-        }
-        catch (Exception ex)
-        {
-            logger.LogWarning(ex, "Не удалось определить кодек видео через ffprobe: {FilePath}", filePath);
-            return null;
         }
     }
